+++ scalus-core/shared/src/main/scala/scalus/compiler/intrinsics/ReprTag.scala
@@ -1,13 +1,5 @@
 package scalus.compiler.intrinsics
 
-/** @deprecated Use [[scalus.compiler.UplcRepresentation]] instead. */
-@deprecated("Use scalus.compiler.UplcRepresentation instead", "0.16.0")
-type ReprTag = scalus.compiler.UplcRepresentation
-
-/** @deprecated Use [[scalus.compiler.UplcRepresentation]] instead. */
-@deprecated("Use scalus.compiler.UplcRepresentation instead", "0.16.0")
-val ReprTag = scalus.compiler.UplcRepresentation
-
 object ReprTagConvert {
 
     import scalus.compiler.UplcRepresentation